feat: remove no confusing void type options & move to nursery
