--- crates/rome_js_analyze/tests/specs/complexity/noConfusingVoidType/invalidAllowAsThisParameter.options.json
@@ -1,15 +1,0 @@
-{
-	"$schema": "../../../../../../packages/@biomejs/biome/configuration_schema.json",
-	"linter": {
-		"rules": {
-			"complexity": {
-				"noConfusingVoidType": {
-					"level": "error",
-					"options": {
-						"allowAsThisParameter": false
-					}
-				}
-			}
-		}
-	}
-}
