--- crates/rome_js_analyze/tests/specs/complexity/noConfusingVoidType/invalidAllowAsThisParameter.ts
@@ -1,1 +1,0 @@
-function doSomething(this: void) {}
