--- crates/rome_js_analyze/tests/specs/complexity/noConfusingVoidType/valid.ts
@@ -1,1 +1,0 @@
-function Foo(): void {}
