--- crates/rome_js_analyze/tests/specs/complexity/noConfusingVoidType/validAllowAsThisParameter.ts
@@ -1,1 +1,0 @@
-function doSomething(this: void) {}
