--- crates/rome_js_analyze/tests/specs/complexity/noConfusingVoidType/validAllowInGenericTypeArguments.ts
@@ -1,5 +1,0 @@
-function printArg<T = void>(arg: T) {}
-logAndReturn<void>(undefined);
-
-let voidPromise: Promise<void> = new Promise<void>(() => { });
-let voidMap: Map<string, void> = new Map<string, void>();
