+++ crates/rome_js_analyze/tests/specs/nursery/noConfusingVoidType/invalid.ts
