+++ website/src/content/docs/linter/rules/index.mdx
@@ -57,8 +57,6 @@ Ensure that the attribute passed to the <code>lang</code> attribute is a correct
 ## Complexity
 
 Rules that focus on inspecting complex code that could be simplified.
-### [noConfusingVoidType](/linter/rules/no-confusing-void-type)
-Disallow void type outside of generic or return types.
 ### [noExtraBooleanCast](/linter/rules/no-extra-boolean-cast)
 Disallow unnecessary boolean casts
 ### [noForEach](/linter/rules/no-for-each)
@@ -316,6 +314,8 @@ Enforce that elements that do not support ARIA roles, states, and properties do 
 Disallow primitive type aliases and misleading types.
 ### [noConfusingArrow](/linter/rules/no-confusing-arrow)
 Disallow arrow functions where they could be confused with comparisons.
+### [noConfusingVoidType](/linter/rules/no-confusing-void-type)
+Disallow void type outside of generic or return types.
 ### [noConstantCondition](/linter/rules/no-constant-condition)
 Disallow constant expressions in conditions
 ### [noControlCharactersInRegex](/linter/rules/no-control-characters-in-regex)
